Changed power from int to double

+++ src/main/java/projlab/ICollide.java
@@ -2,7 +2,7 @@ package projlab;
 
 //TODO: Add Javadoc
 public interface ICollide {
-    boolean collide(Player player, Direction direction, int power);
+    boolean collide(Player player, Direction direction, double power);
 
-    boolean collide(Box box, Direction direction, int power);
+    boolean collide(Box box, Direction direction, double power);
 }
